AG-13306 Rename zoomDOMPRoxy.ts -> zoomDOMProxy.ts
