+++ packages/ag-charts-enterprise/src/features/zoom/zoomDOMProxy.ts
